LayoutTests/fast/images/async-image-large-animated-disabled-expected.txt

This tests large animated images are decoded synchronously unless AnimatedImageAsyncDecoding is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.imageDecodeCount(image) is 0
PASS successfullyParsed is true

TEST COMPLETE
